requirements.txt
torch
tensorflow-cpu
fastapi
uvicorn[standard]
httpx
beautifulsoup4
python-dotenv
gTTS
openai-whisper
deepface
tf_keras
google-cloud-storage
python-multipart
google-auth
transformers 
huggingface_hub
redis
langchain
langchain-google-genai
sentence-transformers
faiss-cpu
langchain-community
langchain_google_genai